[575883] Create correct connection type and error endpoints
If both ends of a connection are missing this fix now ensures that the
correct types for the connection end point and the connection are used
(i.e., event, adapter and data).

This is handled for connection parseing as well as for updating the
type.

Furthermore the error marker ui shows the correct connection color pin.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=575883
Change-Id: I5cba0ad4d6a0effbd28270293264d84c15f57a32