[575918] Bind connection dragging to canvas size
To make connection dragging more smooth and ensure that the interface
bar is not moved during connection dragging the connection and points
are always bound to the canvas.

On reconnection the routing data is updated so that all pins can be
reached and the canvas does not need to grow. If the redragging is
abandoned or reconnected to the original pin the original routing data
is restored.

As part of that the abstractReconnectConnectionCommand has been cleaned.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=575918
Change-Id: I55c92fe6a3838d5936fe270947d9ac7cf34896ed