[576073] Make systemmanager and system editor handling deleted projects
With this fix the systemmanager and the system eitor handle more
gracefully if a user deletes a project from disk while at the same time
editors of the project are open in the background which have not been
loaded yet.

To test the fix the following steps are needed:
 1. Import or create a new project
 2. Open several editors of the new project
 3. Open an editor of another project and keep it as the current active
one
 4. Close 4diac IDE
 5. Re-open 4diac IDE
 6. Without selecting any of the editors delete the project created in
step 1. Check delete from disk
 7. Project should be cleanly removed, all editors associated with that
project should be removed

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=576073
Change-Id: I63ae5aa5c00db7f8311263d6c9ab356164d3514b