[576677] Only use pin name for subapp pin generation
With this change now when fbs are added to subapps and subapp pins need
to be created now only the target pin name without fb instance name is
used. This reduces the size of subapp pin names.

In case of duplicated names the CreateInterfaceElementCommand ensures a
unique and correct name. This is the same code as when a new pin is
crated manually.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=576677
Change-Id: I87f269cb3266209e2d97754a3f4acaf800a1f82e