[575041] Added content adapter to MonitoringEdit Part
The MonitoringEdit part now listens to the deletion of its interface
Element. This also removes them from the watches view.

The listener is now added to all networks in the subapp hierachy. This
is necessary, otherwise a deletion of a nested subapp would not be
recognized if the subapp hierachy has more then two levels.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=575041
Change-Id: I6499797c0a22e41deb0660aa57e082b94bb0e5d5