[576876] Implement Hidden Connection in Model
Invented a visible attribute in the xml as child for the connection to
save the visible state of it.
Also an entry for the property view has been added, so it can be
deactivated.
Activation is currently only possible in the XML, as the connection is
not clickable any more.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=576931
Change-Id: I0c2ce5712bf216baa73c67b882c6fd2eb63a6753