[576118] Fixed Exception when editing subapp interfaces internally
This fixes a wrong refresh visuals in the EditorWithinterfaceEditPart
that lead to a double bring down of the DirectEditManger for direct
editing interface names in the interface bar.

Furthermore this fix cleans up a bit the LabelDirectEditManager and by
that fixes the short cut actions for copy, cut and paste.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=576118
Change-Id: I13561d22ef53a0fce0b484a38e24dad747390662