[577485] Fixes validation of ANY literals
Literals of various ANY types have not been validated, as they were not
part of the instanceof casace. The solution is now to mandatory demand a
type specifier if a literal for an ANY input shall be entered.
With that the literal gets clearly typified and will be checked
accordingly.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=577485

Change-Id: Idff317ea6387d072a9c539d2249e97c1f97cc630
Signed-off-by: Martin Melik Merkumians <melik-merkumians@acin.tuwien.ac.at>