[577486] Fixes validation of DATE
DATE types could not be validates, as the string is checked for its type
and then compared if it is applicable to the input pin.
For the DATE data type there was a bug in
ElementaryDataTypeCompatability, not marking it to be compatible with
itself.
This is fixed in this commit.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=577486

Change-Id: If6933c9590f538235043b25120934024db23ac69
Signed-off-by: Martin Melik Merkumians <melik-merkumians@acin.tuwien.ac.at>