[577502] First version of saving groups to XML
Groups are added like FB instances with their own tag and without type.
FBNetwork elements that are part of a group have an attribute with the
name of the group. This design should ensure a better compatibility with
other IEC 61499 tools. As they can drop the group and still have the FBs
and all their connections.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=577502
Change-Id: I244c86754d082240a63be7034a821a62bac1c6ee