[577501] Interface EditParts should not cache referenced value EPs
The interface edit part cached the value edit part once it found it.
While this is convenient it can lead to problems when the interface
Editpart or its FB is moved to other containers. Then invoking direct
edit from the interface element can lead to NPEs.

This happened especially when adding FBs to groups via drag and drop.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=577501
Change-Id: I0cfe6f8384959196ae8c0f0ecd7041c8b125395f