[577788] palette entry model change led to types without palette entry
With the change for handling editable type models separated from the
type models used for instances a severe issue was introduced. the two
different types (type and typeeditable) where added with a containment
relation into the palette entry. This seemed to reduce effort in
managing types and palette entries.

However the opposite is the case. Everytime the edited type is saved it
will update the non editable type. By that all existing type instances
will loose their containment relation to the palette entry. This not
only makes the instances non updateable also it can lead to NPEs in
several places.

This fix now reverts the original change by manually setting the palette
entry property of library elements

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=577788
Change-Id: I15bcf595871a5a4e3430c8bb3cbdb645d2c4c0e5