[533011] improved apreance of fb network in dark theme
First the shadow border now uses the forground color which is correctly
set in light and drak theme.

Second for the editors background either in the prefernece store or in
the theme manager is searched to get the right color. The first is
needed when not the default theme is set. The second for when the
default is set.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=533011
Change-Id: I7c24f4107eb63a1f17daa6e773fbd443a2c6e2f8