[578264] Removed complete update of subapp interface pins
In order to reduce refreshes of networks inside subapps the full removal
and add of subapp pins done in the refreshvisual has been removed. The
original purpose was to update the interface pin displacement after
autolayout. This is now handled in dedicated adapters of the pin
editparts.

In edition the interfacename virtual model element is now cached by
their respective editparts this also reduces refreshes and garbage
creation.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=578264
Change-Id: Iaed568b3848e33b800a10fd49da384efd6350226