[578299] Changes to literal rules
As data types for variables shall be taken from the data type library,
the rules for literals need to be changes to make the type specifiers
distinguishable from the types (e.g. BOOL vs BOOL#). Lexer tokens are
needed to give the lexer/parser enough information to distingush the
different literal variants.

Bug: https://git.eclipse.org/r/c/4diac/org.eclipse.4diac.ide/+/189825/2

Change-Id: I5b6ad42c6be89aa3609274d9324f3d695b72d403
Signed-off-by: Martin Melik Merkumians <melik-merkumians@acin.tuwien.ac.at>