[578284] correctly create the initial size of groups
With this commit now on creating FBs from existing FBs the required
bounds of the FBs are considered for the initial size of the group.

Furthermore as needed for this the positioning of FBs when dragging them
into and out of groups is updated to the new group graphical model.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=578284
Change-Id: I289798490b687e2589142ecadf9ce28aee8f3419