[578610] New reusable pin-info property sheet
Extracted the table for in/out connections into a separate widget and
altered the pin info widgets.

Refactored refresh(), removed some comments, did minor alterations to
PinInfo classes, added a new extension point and moved the PinInfo
widgets to fordiac.gef plug-in.

Changed plugin.xml of the application for the extension.

Added a propertyTab in the application's .xml and added a new method for
composite creation in PinInfoSection.

Separated PinInfoSection into 2 sections and added filters for them.

Removed duplicated code from sections.

Added a filter for pin tabs and viewer for in/out connections (still
needs work).

Fixed typeSelectionWidget to render the types again.

Deleted old code in the InterfaceElementSection which was causing build
failure.

Fixed connection deletion button.

Fixed the header in InternalConnectionsViewer.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=578610
Change-Id: Ieeee09acb93abcf0d388af5256a05a8701a83e1d
Signed-off-by: dunjazivotin <dunja.zivotin@jku.at>