Bug 578641 Fixes NPE for FOR in ELSIF
When a FOR loop was placed in an ELSIF a NPE was thrown, similar to what
was previously fixed for ELSE paths. The same fix is applied.
When in an ELSIF path, then the eContainer is checked, not the
eContainer of the eContainer due to differences in the AST generation
for IF bodies and ELSE and ELSIF bodies

Conflicts:
	plugins/org.eclipse.fordiac.ide.export.forte_ng/src/org/eclipse/fordiac/ide/export/forte_ng/st/STAlgorithmFilter.xtend

Change-Id: I66a677d979c22a74adb9867feee108faabad209e
Signed-off-by: Martin Melik Merkumians <melik-merkumians@acin.tuwien.ac.at>