Bug 578802 - Fix more spurious validation errors with new ST parser
Fixes more spurious validation errors with the new ST parser. If the STFunction parser was loaded before the STAlgorithm parser,
it also registered validators for the library element types with EObject validation enabled (incl. dangling reference errors).
This suppresses EObject validation also for the STFunction parser to remedy this issue.

Change-Id: Ica7dc645827a8824b89d7b03c11da7684a5e5aa1
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>