Bug 578879 - fixed consecutive add removes of the same type
This fix will cache any add for existing type files until the end of the
change notification. By this later arriving removes will not remove from
the typelib but update the typelib to the new location.

Furthermore in both cases any open editor with that file will be closed.

Some clean-ups in the resource change listener and a safety check in the
type label provider where added as well.

Change-Id: I13061a9ff0fd2ad21a034f782bf7311c278ee5f9