Bug 579111 - Harmonized context menu for group content
Group content is showing the same context menu as the group itself. This
is mainly achieved by using the group as default selection of the
marquee select drag tracker used inside of the group.

Some further safety measures to ensure correct handling of group content
have been added. These are that group and group content are adapting to
groups as well as improved plugin.xml utilizing this feature.

Change-Id: Ie3f8b14a732f1b0b7db3c46dbf564af356beb34f