Bug 579195 - Removed 4diac debug perspective
In order to harmonize debugging the eclipse debug perspective is used
for all debugging purposes. Therefore in addition to removing the 4diac
debug perspective also the system explorer and the watch view are now
shown as part of the eclipse debug persepective.

Some unused icons where removed as well.

Conflicts:
	plugins/org.eclipse.fordiac.ide.systemmanagement.ui/plugin.xml

Change-Id: I1aa195d98f778e1f473f252366615938e3f03f6c