Bug 579272 - Ported internal elements property sheets to filters
With this change the internal var and interanl fb property sheets are
moved to a filter which is independant from the edit parts. Therefore
they will work with the ECC and text editors alike. Also duplicated code
could be removed as now one property sheet can be used for both basic
and simple fbs.

Change-Id: I4c1867b048616e7c82e2695c38f88f7f77b6a03c