Bug 579404 - Continuous synchronization between graphical and ST algorithm editors
Synchronize changes continuously between graphical and ST algorithm editors and not just on save.

In the direction of the library element model to the ST editor, changes are synchronized by a background job that updates the FB type in the ST algorithm resource and relinks the Xtext model to re-resolve cross references that might have changed. The job is triggered asynchronously by a recursive EMF adapter on the FB type listening on changes in the library element model. This has the advantage of keeping the the graphical editors responsive and allowing for the job to block until such a time that is convenient for applying the changes to the Xtext document and model. The job further uses a small delay so as not to synchronize on every minor change.

In the direction of the ST editor to the library element model, changes are synchronized by extending the reconciling strategy of the Xtext editor to also reconcile the ST model with the FB type. This has the advantage of neatly integrating with existing functionality provided by Xtext, which also performs reconciliation in an asynchronous job so the editor remains responsive.

Change-Id: I3d2f4b7d177d7c21b014e0b2fccc5cb68bcd1cff
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>