[clean-up] Always set the connection fbnetwork to the root of the editor
This clean-up commit is preparational work for improving connection
creation across expanded subapps. In this first step the FBNtwork for
connections is always the the root fbnetowrk of the current editor page.
Which can be best accessed via the root editpart and its adapt method.

This also reduced code in the interface bar interface element editparts.

Change-Id: I5118bc8a078f7007193d54ea7dd4bdb8744cb1d3