Bug 579444 - moved parent checking to interface policy
In order to have also the right parent for expanded subapps and a
correct checking if a connection is possible the parent fbnetwork
checking and handling was moved to the InterfaceElementEditPolicy.

Furthermore this will make the connection create commands better usable
and is a prerequiste for implementing a subapp border crossing
connection creation infrastructure.

Change-Id: I6e3954f73b839568263d6f409cb6dc3717bcabe5