Bug 579485 - fixed changed type for untyped subapps
With the new expanded subapp figure where direct editing will lead to
edit the command the inline change type command was not working anymore
as it was tried to edit the non existing comment.

With this fix in collapsed mode the type is changed and in expanded mode
the comment is edited.

Change-Id: I7e3214c6aee948772fbd97e34a79c7e8be85f112