Bug 579413 - update of file modification timestamp to late
On saving types the time modification timestamp of the paletteentry is
updated after the saved type is updated. This can lead to inconsistent
system state.

Furthermore to protect against unnecessary editor reloads the type
editor reload checks if the type changed and only on a changed type
updates the editor.

Change-Id: I461552c6f8570437a7f7758f4c75cfb95493ca47