Bug 579570 - changed group checking location for change type
Especially when an error marker FB was located inside a group and the
type is changed to the an existing one the new FB was not put into that
group. The reason is that the group checks are done to late in the
update cycle. At that point the error marker is already deleted and any
group information lost. By doing this now earlier in the execute and
redo the group is savely set.

Change-Id: Ia008c23ad63577dabfc03ec3f2742ebee1bde162