Bug 579450 - Only write markers if there are markers
When parsing type files the common element create is starting a
workspace job for writing the markers in the background. However this is
also done even when there are now markers to create. The performance of
that can be greatly improved if this job is only created when there are
markers to write.

Change-Id: I9a0b1aa8bb4161d4f398a6252be06c6ca9f51c9f