[579792] Add an exact matching search
Added a small check to provide the user with the option of looking for
an exact FB/pin/etc. since they might know the name and don't want to
see other partial hits.

Also changed the name of the search.

Addressed the get/set comment (they were not needed indeed).

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=579792
Change-Id: Ia8737dedda9bc280592897f3fa590638f32541c8
Signed-off-by: dunjazivotin <dunja.zivotin@jku.at>