[579682] Allow to delete error marker pins
Up to now it was only possible to get rid of error marker pins when all
connections leading to it or originating from it where removed. With
this change now it is possible to explicitly delete an errormarker pin.

Furthermore the interface refresh code was harmonized between subapps
and fbs.

Change-Id: I270b78cb5a6631e5cadb2c45f66d1e9cd0b96ab9