[579918] Editing of pin comments was moved into a separate property tab
A new tab called "Comments" was implemented and within it it's possible
to change the initial values of data pins and their comments.

The failing tests now pass but it seems like it was a problem on my end.

Replaced FBImpl with a more general FBNetworkElement and removed the
SubApp check.

Added saving of the instance comment and a null check for value;
corrected the copyright headers.

Added commands and changed the superclass. Added the command for name
saving as well.

Fixed the text which appeared when value was null for data outputs of a
subapp.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=579918
Change-Id: Ie139b29a3ea1912101e0abb8ac539297a49cdaba
Signed-off-by: dunjazivotin <dunja.zivotin@jku.at>