Bug 579926 - Create Save as dialogue for struct editor
Implemented wizard page and enabled save as in the file menu.

Used the wizard factory to create the save as struct type wizard.
It is now possible to perform File/Save as inside the STRUCT Editor. New
typed is opened in a new editor.

Saved the dirty state of the editor. You do not need to save the
parent file which you want to save as a new file

For overwriting the current editor is saved. It is now possible
to either overwrite the existing type, or the window with the new
type is loaded.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=579926
Change-Id: I50796414e60ee7a2a694480bd15faae2b26cb302