Bug 579939 - Fixed ExportSatusMessageDialog's layouting
As the layouting was not correctly configured the text field grew very
large in case of many errors. With the fix the text field is correctly
resized and if to small correctly updates the scroll bars.

Furthermore the warnings and errors have been swaped and the
warning/error messages are now indented and prefixed with a dash.

Change-Id: I247bf224f294093f6849f3eaee7f03466d237671