Bug 579954 - Show the selected table cell highlighted
Changed the default table configuration such that always the focus cell
is highlighted. With the arrow keys the selection can be moved around.
Enter or F2 will enable the editing of the cell if it is editable.
Insert will insert a new row below, delete will delete the row of the
selected cell.

Change-Id: I1ad6da44b54ccf8a1c2805b471f6a75ad6e73876