Bug 580024 - deactivate mapping check for connection creation in mapping
The Map to command tries to create all necessary connections in the
resource. It uses the create connection command which per default tries
to mirror the connections of mapped entities. In the case of the mapping
this resulted in creating a connection in the source of the mapping,
which lead to an unnecessary error pop up.

This patch deactivates the mapping check for mirrored connections as it
is not needed during the mapping.

Change-Id: I93712e28f3448f82aa0ee09c5c7176412d0a38ec