Bug 580024 - reworked opposite connection creation checking
The opposite connection creation checking performed a full check as it
would create a new connection. At that point we know that in principle
the connection should be valid, otherwise the source connection would
also not be possible. While this is a small performance point it has a
bigger issue that on reconnect unnecessary and distracting error pop-ups
occur. With this change these are now gone.

Change-Id: Idecc1a0c016a4f46828447f522c6ad00de742b9f