Bug 580034 - Harmonized Subapp Content and Group Content EPs
With this change the subapp content area and the group content area are
behaving more equal. Especially there is no offset and automatically
size calculation.

It is now also possible to use direct edit for creating FBs in the
subapp content area.

Furthermore duplicated code for creating FBs has been harmonized in a
factory method provided by the AbstractCreateFBNetworkElementCommand.

Change-Id: Icf21e6950da05b8fa9ef27b402ac650c0c101527