Bug 580048 - Fix connection type constraints with generic source
Fix the connection type constraints when the source has a generic type. In this case, the source adapts to the type of the destination. This means the type of the destination has to fit in the (generic) type of the source instead of vise-versa, which would be the case normally.

Change-Id: Ia697d2e7b4d6a14dffd6f2aa7fcbaa56bb36b405
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>