Bug 580064 - Fix Xtext commands not available in ST algorithm editor
Fix Xtext commands not available in context menu and shortcuts in the ST algorithm editor. The adapt command filter tries to resolve the adapter class using the bundle of the original object, which fails for the XtextEditor if it is not in the dependencies of the FB type editor bundle.

Change-Id: I2d24c61d561e5c91b02bfed0b6886321059ab977
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>