Bug 580034 - fixed dragging on and of collapsed subapps
The accidently removed FBAddtoSubappLayoutEditPolicy has been readded
and activated when the subapp is collapsed. This policy is in charge of
handling drag of fbs on collapsed subapps.

Furthermore this changes fixes that dragging subapp instance names led
to class cast exception and that dragging fbs on top of expanded subapps
resulted on code freezes.

Draging fbs on expanded subapps is still not working.

Change-Id: Idc59c4b2c3672af9c23b0914107373946042f919