Bug 580184 - moved ConfigureableObject out of the inheritance hierarchy
In order to allow more objects to have attributes ConfigureableObject
was moved to an independent object. All current classes which where
using it have it now as second inheritance.

Added ConfigurableObject to TypedObject so that all types can have now
attributes.

Used this to fully recreate the model and the model.edit.

Change-Id: I89fe902ee26b06aa4e8351fcc9b95533b4ecbe1c