Bug 580252 - Fix remove stale persisted markers on editor change
Fix remove stale persisted markers when the problem is changed/fixed in the editor. Model annotations are only deleted if they have the annotation type corresponding to the Xtext editor. In case of marker annotations, the annotation type is inferred from the marker type and its super types. Due to the search order, the resource problem marker was found before the emf diagnostic marker, which resulted in the wrong annotation type being used.

Change-Id: Id3eaf796bd46061e9fd8fb5102a93b6eefd50190
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>