Bug 580268 - Ensure that the con property sheet only has valid cons
In rare cases the property sheets can be updated with half correct
models. For example during deletion when parts of the model elements are
set and the viewer is refreshed. This can lead to strange situations for
the property sheets. With this fix now the connection property sheet
only accepts fully valid connections.

Change-Id: I7ad8beefad39111ba1f686a0a054ab340d026ebc