Bug 579444 - moved compatibility check to subapp crossing command
Subapp app crossing connection validity is now checked in the can
execute method of the command. There source and destination are correct
which makes it more correct.

Furthermore a double adding of the subappnetwork of the source and
destination pin was removed.

Also the isSwap needed had didn't use the correct network.

Still open are connections that have the first or last segment as a
subappcrossing connection (i.e., subapp input to the same subapp output
and vice versa),

Change-Id: I59b016060904a425579cd9a8d1ab61fa10e7317d