Bug 580338 - trim handler fixed for empty content
Empty content resulted in trimmed element moving to (0,0). Now the
location is fixed. Also the comment size is now taken into account to
ensure that the comment is not cut off. Keeping the width of the comment
intact will also ensure that the formatting is not destroyed.
Previously, the group size was too small when the comment ended up
needing additional lines. Removed margins from expanded subapp comment
figure to match group comment figure.

Change-Id: I9d3d5698bb1af5e32914ba0d51d03c265bea07ce