Bug 580034 - set default position in subapp to 0/0
With the changes made to 4diac IDE in the past (especially sidebar) the
offset used for the new subapp calculation did not make sense anymore.
Earlier, the offset was introduced for a better-looking layout with the
fixed interface pins (layouting of the connections). The recent changes
have affected the positioning of blocks greatly. This is a first step
towards fixing position-related bugs.

Change-Id: I6c9e53b8a1a2926c5622f8052d73576cb555b1b8