Bug 580055 - adjust position for convert to group
Adjusted flattensubappcommand because the network
elements were selected as part of this command. Resulted in annoying
movement of the editor and incorrect selection visualization between the
updates could be avoided. Fixed undo/redo selection visualization. Still
open is a small offset in margins and because of the subapp instance
name.

Change-Id: I8f337135d07e596af1185e32135434e8d4dfb2e6