Bug 580055 - consider instancename and pins for subapp <--> group
The instance name affects the size of the expanded subapp contents and
has to be considered. Previously, the bottom of FBs was cut off when a
trimmed group was converted to a subapp. As a group does not have an
instance name, a default height is assumed (measured from the default
diagram font in 4diac IDE). Also harmonized margins between subapp and
group.

The pins take some space as well. The only reasonable option is to
perform a trim afterwards if necessary.

Change-Id: Ia81feae10e850f650deb78c9a0c3753dae546b82