Bug 580385 - Fixed delete issues of values in groups and subapps
The position correction of values shown in groups or expanded subapps
was done in the setLayoutconstraints of the parent editpart. The problem
is that this method is also called on reordering of children (e.g.,
deletion of a child) Therefore on every deletion an offset was added.

With this fix the position compensation needed for values in groups or
expanded subapps is performed in the editpart itself. This should solve
at least a few of the problems.

Change-Id: I2ce84334242005785e0aae89cf1db310071defcf