Bug 580415 - reworked connection source and destination check
With this change the source and destinations of connections are
individually checked. This fixes the problem of allowing wrong
connections inside subapps and CFBs. Furthermore, I hope this code is
better readable and maintainable.

With this change it was also necessary to rework the canExecute method
for the createsubappcrossing command. There with the new checks the
problematic errornouse connections are now also not allowed anymore.
This will only work when we find a solution for the last segment.

Change-Id: I335e8a2935e3afc3fc4682b5c3be75d8b11a3f7f