Bug 580416 - Added null check for co-located connection finder
The co-located connection finder is using the editpart registry for
finding editparts for models. However it can be that in some, mostly
error, cases no editpart is found for a model element. In this case a
null check needs to be performed. This is added with this commit. Also
the for loop is changed to a stream.

Change-Id: I5fb5c051832fe977aab4a37f91333bb12aebd4cb