Bug 580488 - Deployment console shows only error messages
With this change the deployment console only will print requests and
responses when there is an error response. This greatly reduces the
number of messages print and therefore reduces the resource needed.

In addition to give the user some indication on the deployment process
the connecting and disconnection from a device are printed as comments
to the deployment console.

Change-Id: Ie819d3902bc0a1aa297dc3f640c10c4ea8c23582