[579702] add input var list to transactions
When executing event sequences, the variables have to be set manually to
the FB type enclosed in a respective runtime. However, the FBtype is
altered after every input event and therefore running a full event
sequence at once is not feasible. To improve the API of the interpreter,
FB transactions can now provide a list of input variables that should be
updated right before executing an FB based on an input event occurrence.

Change-Id: Ib7a64d8a63707b6e35ded22554f356e4e43ca769