Bug 580598 - reworked delete handler for monitoring elements
The delete handler checks now more explicit the different cases when a
monitoring element should be removed. This should correctly handle also
parent group and parent subapp deletions.

Furthermore I fixed the deletion of interface pins of subapps. In this
case now also the delete handler is invoked.

What is not yet working is that the ui is correctly cleaned. In case of
deletion there can be still monitoring figures left. This will be
handled in a separate commit.

Change-Id: I4d94f7a72b14239847ed94e425ef31d148d21190