Bug 580531 - store random data in new transaction variables
Previously, only the first event of a sequence could get random data, as
the fbtype copies deleted all other information. With the new patch
adding inputVariables to transactions, the random generation now adds
the variables to the transactions (keeping the With between events and
data in mind).

RecordServiceSeq now works on a copy to avoid editing the type in the
interpreter (major bug).

Change-Id: I04071a322243bc7a5bf84d53ca96169088fd027f