Bug 580671 - Subapps created from fbs in a group created in group
With this fix subapps that are created from a set of fbs located in a
group will be placed in the group the fbs have been.

Furthermore more enablement checks are added to the creation of groups
and subapps. If they are created from a list of fbs they can only be
created if all fbs reside in the same fbnetwork. Subapps only of the fbs
are all in now group or all in the same group.

Change-Id: Ic97fde725c604f1b55532fe6fcc4c055335bde8e