Bug 580722 - Used ResizeGroupOrSubappCommand for adding elements
Adding elements to groups or expanded subapps used an own resizing code.
This code did not consider hierarchically nested expanded containers.
Also it didn't wait for the viewer to update.

By replacing this with the ResizeGroupOrSubappCommand not only the
nested resize is fixed also the strange problem that expanded subapps
could grow far to the left when new interface pins where created.

Change-Id: I21e33cb5cb6a3f2e005db149f75a01e378b4ae97