Bug 580722 - Fixed positioning of add to subapp
Add to subapp was setting the position at two different places. This was
especially problematic when new interface elements where created, as
there a resize check is performed.

This change moves all position settings of each added FBNetworkElement
to a single command which is executed before any connections are
created. With that strange resize problems that occurred are fixed.

Furthermore a performance issue in the ResizeGroupOrSubappCommand was
also fixed.

Change-Id: If02349d6425be13b5947db724e1236833c67942f
Signed-off-by: Fabio Gandolfi <fabio.gandolfi@jku.at>