[bugfix] Fixed minor bugs for error marker handling and update type
The AbstractUpdateFBNElement command now uses its own deleteconnection
command, as a reconnected deleted the error marker on the opposite block
which is not updated.
This leads to the result that the connection opposite is not existing
any more.

Furthermore the assignable check for events have been adapted, since the
old code was wrong and wasnt even capable of getting executed

Change-Id: I35514e9b68a0e1c343867b39818aa21b787b4908