Bug 580872 - fixed update refresh problem in documentation editor
The richtext editor sends a lot of modifcation notifications. Even when
the content is not changed. The only solution I found is to check if the
content is different from what we have in our model.

When updating to store the documentation differently the check for
modification was not fixed as well.

Change-Id: Iead384d9b28f6c69747dff6ec0fb797d8acdad96