Bug 580987 - add editor for forcing arrays of structs
add support for struct force dialog for arrays of structs. The can also
be forced. For a better visualisation I added names, which are the
indexnumber of the element in the array.
Since they are arrays all element have the same type and if you force
any type always the first element of the array will be changed, thats a
problem with the parser. Bugreport already exist.

Change-Id: I7cd81c784b6a5a14e142646dfe82321b9515198d
Signed-off-by: Fabio Gandolfi <fabio.gandolfi@jku.at>