Bug 580605 - Provide dedicated link helper for 4diac specific behavior
This fix adds a notification when any of the multi-page editors changes
the tab so that the link with editor action of the common navigator
framework is triggered to update the viewer.

In addition a dedicated link helper for FBNetworks is added which will
return the right selection to react to the tabs.

This new link helper is also used as replacement for the resource link
helper from the Eclipse platform. With that now when we drag type files
for creating elements the tab is not switched any more.

Change-Id: If15f8bafa374ed2845454f008a95418dd75abde2