Bug 581067 - reworked FBNetwork link helper to a system explorer version
The FBNetwork link helper is renamed to SystemExplorerLinkHelper and is
now handling all editors we typically have in a 4diac IDE project.

Furthermore the SystemContantProvider, which is the main content
provider in the System explorer is now correctly calculating the parents
for FBType content. The reason for that is that the
SystemContentProvider will be asked first and only for that task as it
is registered for any EMF content.

Change-Id: I035ef28f7c9dd3cd7f358d311554f04696685c48