Bug 581100 - centralised watched struct editing
Editing watched structs in the watches view previously changed the
string value in the monitoring element which in turn updated the watch
value nodes. The heavy reliance on string manipulation resulted in
unexpected behaviour with nested structs. With this commit edits are
done on the nodes which are then parsed into the full string literal of
the struct.

Change-Id: I89ea4d81b17368b8e3425cb62faf891001f12436