Bug 581101 - added editor ID to open editor to get the right editor
Open editor in the EditorHelper used the simple eclipse open editor
method with input and editorId. However in this version the editorId is
not checked and if an editor with the same editor input (eg., a text
editor) was already open this was given and not the right editor type.

Change-Id: I4e0158dd1bc19df800baec4cb57f8f027960c1cb