Bug 581138 - Size of "Initial value" text label too large
When there are many elements in a struct which is a pin in a FB, the
size of the "Initial value" text field was too large and displayed all
of the initial values. Now a width hint has been added to the layout
which should allow only a certain part of the text to be shown. The hint
has been moved up in the hierarchy so that all property sections with a
similar structure would be affected.

Change-Id: I936850a77bca302d191992a229c1b2df3d5fbfca
Signed-off-by: Dunja Zivotin <dunja.zivotin@jku.at>