Bug 580639 - Selection of the NatTable check box with the mouse
Until now we could only check/uncheck the boxes in the "Visibility"
column by selecting the cell and pressing space bar. With this small
addition, the check boxes can be changed with the left click of the
mouse. Important: the user should click on the actual box, not just the
cell containing it.

Change-Id: Ib3386147e1577a479f61600975270b5c1457b9f2
Signed-off-by: Dunja Zivotin <dunja.zivotin@jku.at>