Bug 581182 - synchronize type saving and timestamp updates
To always have a consistent type entry on save now the writing to the
file and updating the modification timestamp is synchronized with also
getting the modification timestamp.

This should avoid any resource change listener to kick in in the middle.

Furthermore the typeentry is further protected with synchronization and
also the saving is handled via the type entry.

With that also a further harmonization step between type and system
handling.

Change-Id: I8e2e0d1806f766bac0f7dfb9cf756b26d997c1e9