Bug 581006 - Fixed layout data generation of expanded subapp
The layout data generation of expanded subapps was adding the target
connection of the subapp input pins to thelist of connections. As the
source of these connections is not part of the layout data list an NPE
was thrown. Also this shouldn't be done as expanded subapp layouting
should only layout internal subapp data.

Change-Id: I8dc6d4d03c243d4edb7da1760e297fad19a85403