Bug 581675 - Fix duplicate type saving in ST algorithm editor
The STAlgorithmDocumentProvider is not only updating the model but also
saving the full type. However, this is supposed to be done once by the
save method of the FBTypeEditor for all sub editors. As such, this can
lead to duplicate saving of the FB type. This change detects if the
document is opened directly from a file (e.g., for refactoring) or
within an FB type editor, which will skip the duplicate save. It also
adds IFileEditorInput as superinterface for FBTypeEditorInput in order
to simplify handling in the ST algorithm editor.

Change-Id: I3f3de7dfc2cee323cb499b30699188fe717b070d
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>