Bug 581307 - Fixed selection handling in debug viewer
As the debug view didn't use an edit domain no default tool, which is
typically the selection tool was applied to the viewer.

Furthermore the ZeroOffestFreeformCanvas didn't calculated the client
area correctly which resulted in a dead area on the left side where
selection is not working.

Finaly to give selection feedback the value editparts are given hover
and selection feedback. Any other element is not selectable.

Change-Id: If2a60a2c04ce39298e6b0709cb9f0a3a31ee6a93