Bug 581310 - Fix dependency resolution for ST code in forte_ng exporter
Fix dependency resolution for ST code in the forte_ng exporter by
properly distinguishing between dependencies from and to a definition.

The dependency resolution provides a list of dependencies from objects
in the current source to objects defined elsewhere, such as from the
declaration of a variable to its data type. This is then later used to
determine the necessary include statements. However, there is also the
case when a feature expression has an external dependency, such as to a
function. This was previously handled recursively through the same
method. As such, the direction of the dependency as seen by the method
is reversed, now handling an incoming dependency to the object in
question. Besides being quite confusing, this also causes ambiguity
between each direction. This commit provides a fix that separates both
cases.

Change-Id: Ie04a3e5143d59d5e0bdd66f034655df70036431f
Signed-off-by: Martin Jobst <martin.jobst@posteo.de>